refactor: Create new OAuth callback controller and actions for Azure and deprecate the old controller
