+++ alert-common/src/main/java/com/synopsys/integration/alert/common/rest/AlertRestConstants.java
@@ -16,6 +16,7 @@ public final class AlertRestConstants {
     public static final String OAUTH_CALLBACK_PATH = CALLBACKS_PATH + "/" + OAUTH;
     public static final String CONFIGURATION_PATH = AlertRestConstants.BASE_PATH + "/configuration";
     public static final String AZURE_BOARDS_CONFIGURATION_PATH = AlertRestConstants.CONFIGURATION_PATH + "/azure-boards";
+    public static final String AZURE_BOARDS_OAUTH_CALLBACK_PATH = AlertRestConstants.OAUTH_CALLBACK_PATH + "/azure";
     public static final String EMAIL_CONFIGURATION_PATH = AlertRestConstants.CONFIGURATION_PATH + "/email";
     public static final String JIRA_SERVER_CONFIGURATION_PATH = AlertRestConstants.CONFIGURATION_PATH + "/jira_server";
     public static final String SETTINGS_PATH = AlertRestConstants.BASE_PATH + "/settings";
